% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laadan_dictionary_text.R
\docType{data}
\name{Laadan_dictionary_text}
\alias{Laadan_dictionary_text}
\title{Dictionary for the Laadan language}
\format{An object of class \code{character} of length 3383.}
\usage{
Laadan_dictionary_text
}
\description{
A character vector of lines of text scraped from the "official" 
Laadan-to-English dictionary at 
\url{https://laadanlanguage.wordpress.com/dictionaries/laadan-to-english-dictionary/}.
The text has been lightly cleaned by eliminating extra lines in the original that broke the 
patterns of lines alternating between the Laadan word and the English translation of that word. Many of
the definition lines contain an entry in curly braces. This indicates when the word was added to the language.
Words without such entries are contributed by Laadan enthusiasts.
}
\seealso{
\code{\link{Laadan_labels}}
}
\keyword{datasets}

