% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ocean_states.R
\docType{data}
\name{Ocean_states}
\alias{Ocean_states}
\title{US States bordering an ocean}
\format{A data frame with 50 rows and 2 variables}
\source{
Generated by hand by DTK on 2016-01-10
}
\usage{
Ocean_states
}
\description{
For the 50 states, whether or not the state borders an ocean.
Note that Washington, D.C. is not included
}
\details{
\itemize{
  \item state. Name of the state
  \item ocean. "yes" or "no".
}
}
\keyword{datasets}

