% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/State_pop_area.R
\docType{data}
\name{State_pop_area}
\alias{State_pop_area}
\title{Population and area of US States}
\format{A data frame with 51 rows and 3 variables}
\source{
Scraped from \url{https://en.wikipedia.org/wiki/List_of_U.S._states_by_population_density} on 2016-01-10
}
\usage{
State_pop_area
}
\description{
For the 50 states + the District of Columbia, the area of the state and the 
population as estimated by the US Census in 2015
}
\details{
\itemize{
  \item state. Name of the state
  \item pop_2015. Population of the state in 2015.
  \item area. Area of the state in square miles.
}
}
\keyword{datasets}

